+++ CHANGES.rst
@@ -34,6 +34,9 @@ Changelog
 -   Defer Comet and Twisted imports until they are actually needed by the
     VOEvent broker. This should speed up worker startup by about 0.2 seconds.
 
+-   Defer imapclient imports until the email client starts. This should speed
+    up worker startup by about 0.1 seconds.
+
 0.12.3 (2020-03-24)
 -------------------