+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-0.12.4 (unreleased)
+0.13.0 (unreleased)
 -------------------
 
 -   Set FAR threshold for early warning alerts to once per day. With the trials